Multiplex serial port and monitor
This provides another way to log into the guest even when
no display is shown and no SSH daemon is listening. Just
make sure the guest runs a getty on its first serial port.

Use the following command to connect to the serial port:

$ socat -,cfmakeraw,escape=0xf .kitchen/<instance name>.mon

Ctrl-a followed by c switches between the serial port and
the QEMU monitor. Ctrl-o exits socat.